# openapi.yml
openapi: 3.0.3
info:
  title: GPay Wallet System API
  version: 0.1.0
  description: Implemented API contract for the Auth Service, Wallet Service, Payment Service, and Mock Gateway Service.
servers:
  - url: http://localhost:8081
    description: Local Auth Service.
  - url: http://localhost:8082
    description: Local Wallet Service.
  - url: http://localhost:8083
    description: Local Payment Service.
  - url: http://localhost:8084
    description: Local Mock Gateway Service.
...
paths:
...
  /payments/top-up:
    post:
      tags:
        - Payment
      summary: Create a payment top-up transaction.
      description: Creates or replays an authenticated payment top-up transaction using durable PostgreSQL idempotency, then calls Mock Gateway through WebClient. Gateway timeouts leave the transaction PENDING.
      servers:
        - url: http://localhost:8083
          description: Local Payment Service.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentTopUpCreateRequest"
      responses:
        "201":
          description: Top-up transaction was created. Duplicate requests with the same idempotency key and payload return the stored response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentTopUpShowResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /payments/webhook/gateway:
    post:
      tags:
        - Payment
      summary: Receive a signed mock gateway webhook.
      description: Validates X-Gateway-Signature using HMAC_SHA256(secret, timestamp + "." + rawRequestBody) before updating the payment transaction to SUCCESS or FAILED. SUCCESS webhooks enqueue one pending wallet-credit outbox event.
      servers:
        - url: http://localhost:8083
          description: Local Payment Service.
      parameters:
        - $ref: "#/components/parameters/TraceId"
        - name: X-Gateway-Signature
          in: header
          required: true
          description: HMAC SHA-256 signature for timestamp and exact raw JSON body.
          schema:
            type: string
        - name: X-Gateway-Timestamp
          in: header
          required: true
          description: Timestamp included in the signature input.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentGatewayWebhookCreateRequest"
      responses:
        "200":
          description: Gateway webhook was accepted and payment transaction status was updated. SUCCESS creates or reuses pending wallet-credit outbox work.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentGatewayWebhookShowResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      description: Optional request trace id. If omitted, the service generates one, returns it in the X-Trace-Id response header, stores it in MDC logs, and propagates it on service-to-service calls.
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Durable idempotency key for money-moving requests.
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation_error:
              value:
                error: VALIDATION_ERROR
                message: email must be a well-formed email address
            insufficient_balance:
              value:
                error: INSUFFICIENT_BALANCE
                message: Wallet balance is not enough for this transfer
            daily_limit_exceeded:
              value:
                error: DAILY_TRANSFER_LIMIT_EXCEEDED
                message: Daily transfer limit exceeded
            same_wallet_transfer:
              value:
                error: SAME_WALLET_TRANSFER
                message: Sender and receiver wallet must be different
    Conflict:
      description: The request conflicts with an existing resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            email_registered:
              value:
                error: EMAIL_ALREADY_REGISTERED
                message: Email is already registered
            idempotency_key_conflict:
              value:
                error: IDEMPOTENCY_KEY_CONFLICT
                message: Idempotency key was already used with a different request
            payment_transaction_conflict:
              value:
                error: PAYMENT_TRANSACTION_CONFLICT
                message: Payment transaction was already credited with a different request
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: Resource was not found
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: Authentication is required
            trace_id: trace-auth-123
    TooManyRequests:
      description: The authenticated user exceeded the payment top-up rate limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: RATE_LIMIT_EXCEEDED
            message: Maximum 5 payment requests per minute allowed
    ServiceUnavailable:
      description: The service cannot safely process the request because a required dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: RATE_LIMIT_UNAVAILABLE
            message: Payment rate limit cannot be verified
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    InternalTokenAuth:
      type: apiKey
      in: header
      name: X-Internal-Token
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request
        trace_id:
          type: string
          nullable: true
          description: Request trace id generated or preserved by the service when the error is handled by the service exception layer.
          example: trace-payment-123
...
    PaymentTopUpCreateRequest:
      type: object
      required:
        - wallet_id
        - amount
        - gateway_mode
      properties:
        wallet_id:
          type: string
          format: uuid
          description: Wallet intended to receive the top-up credit.
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Top-up amount in whole IDR.
          example: 75000
        gateway_mode:
          type: string
          enum:
            - SUCCESS
            - FAILED
            - TIMEOUT
          description: Mock gateway behavior to request for this top-up simulation.
    PaymentTopUpShowResponse:
      type: object
      required:
        - payment_transaction_id
        - wallet_id
        - amount
        - status
        - created_at
      properties:
        payment_transaction_id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          description: Top-up amount in whole IDR.
          example: 75000
        status:
          type: string
          enum:
            - PENDING
        created_at:
          type: string
          format: date-time
    PaymentGatewayWebhookCreateRequest:
      type: object
      required:
        - payment_transaction_id
        - wallet_id
        - amount
        - status
        - gateway_reference
      properties:
        payment_transaction_id:
          type: string
          format: uuid
        wallet_id:
          type: string
          format: uuid
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Top-up amount in whole IDR.
          example: 75000
        status:
          type: string
          enum:
            - SUCCESS
            - FAILED
        gateway_reference:
          type: string
          example: gw-4f437123-8c2d-41c5-9300-a0f8f72fd6e3
    PaymentGatewayWebhookShowResponse:
      type: object
      required:
        - payment_transaction_id
        - status
      properties:
        payment_transaction_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - SUCCESS
            - FAILED
...